# openapi/okoscope-v1.yaml
openapi: 3.1.0
info:
  title: Okoscope Web API
  version: 1.0.0
  description: Stable API contract for the separately deployed Okoscope Web UI. Timestamps are RFC 3339 UTC values.
servers:
  - url: /
security:
  - sessionAuth: []
paths:
...
  /api/v1/organizations/{organization_id}/runtime-retention:
    parameters: [{ $ref: '#/components/parameters/OrganizationId' }]
    get:
      operationId: getOrganizationRuntimeRetention
      description: Read effective retention settings.
      responses:
        '200':
          {
            description: Retention settings,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/RuntimeRetentionPolicy' } },
              },
          }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
    put:
      operationId: putOrganizationRuntimeRetention
      description: Organization owner required. Changes affect subsequent maintenance batches.
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/RuntimeRetentionPolicy' } } }
      responses:
        '200':
          {
            description: Retention settings,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/RuntimeRetentionPolicy' } },
              },
          }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/projects/{project_id}/runtime-retention:
    parameters: [{ $ref: '#/components/parameters/ProjectId' }]
    get:
      operationId: getProjectRuntimeRetention
      description: Read effective retention settings.
      responses:
        '200':
          {
            description: Retention settings,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/ProjectRuntimeRetention' } },
              },
          }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
    put:
      operationId: putProjectRuntimeRetention
      description: Organization owner required. Changes affect subsequent maintenance batches.
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/RuntimeRetentionPolicy' } } }
      responses:
        '200':
          {
            description: Retention settings,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/ProjectRuntimeRetention' } },
              },
          }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteProjectRuntimeRetention
      description: Organization owner required. Changes affect subsequent maintenance batches.
      responses:
        '200':
          {
            description: Retention settings,
            content:
              {
                application/json:
                  { schema: { $ref: '#/components/schemas/ProjectRuntimeRetention' } },
              },
          }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: okoscope_session
      description: Opaque HttpOnly browser session; clients send requests with credentials enabled.
  parameters:
...
    ProjectId:
      { name: project_id, in: path, required: true, schema: { type: string, format: uuid } }
    OrganizationId:
      { name: organization_id, in: path, required: true, schema: { type: string, format: uuid } }
...
  headers:
    RequestId: { schema: { type: string }, description: Correlation identifier for the request. }
...
  schemas:
...
    RuntimeRetentionPolicy:
      type: object
      additionalProperties: false
      required: [enabled, raw_days, history_days]
      properties:
        enabled: { type: boolean, default: false }
        raw_days: { type: integer, minimum: 1, maximum: 3650, default: 30 }
        history_days:
          type: [integer, 'null']
          minimum: 1
          maximum: 3650
          description: Total observation age in days; must be at least raw_days. Null preserves snapshots forever. Processing uses complete UTC days.
    ProjectRuntimeRetention:
      type: object
      required: [override, effective, inherited, source]
      properties:
        override:
          anyOf:
            - { $ref: '#/components/schemas/RuntimeRetentionPolicy' }
            - { type: 'null' }
        effective: { $ref: '#/components/schemas/RuntimeRetentionPolicy' }
        inherited: { $ref: '#/components/schemas/RuntimeRetentionPolicy' }
        source: { type: string, enum: [organization, project] }
...
    Error:
      type: object
      additionalProperties: false
      required: [error, message, request_id]
      properties:
        error:
          type: string
          enum: [validation_failed, unauthorized, invalid_credentials, email_verification_required, registration_disabled, registration_conflict, untrusted_origin, forbidden, not_found, organization_not_found, project_not_found, application_not_found, user_not_found, user_not_eligible, setup_already_completed, invalid_setup_token, setup_rate_limited, privilege_confirmation_required, current_password_invalid, self_promotion_forbidden, last_super_admin_required, last_organization_owner_required, organization_limit_reached, invitation_unusable, invitation_account_mismatch, invitation_exists, rate_limited, mail_unavailable, idempotency_key_reused, operation_already_completed, internal_error]
        message: { type: string, example: resource not found }
        request_id: { type: string, example: 0ec02ed2-8483-4981-893e-bffc535897d7 }
        fields:
          type: object
          additionalProperties: { type: string }
          description: Present for field-level validation failures.
...
  responses:
...
    Error:
      description: Correlated API error
      headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }